Fix generate_soft_labels: force fp32 to match training dtype

+++ models/deberta-v3-large-nlp-en/generate_soft_labels.py
@@ -100,7 +100,7 @@ def generate(model_dir: str):
     print(f"Device: {device}", flush=True)
     print(f"Loading teacher from {model_dir}...", flush=True)
 
-    model = MultiTaskNLPModel.load(model_dir, encoder_name).to(device)
+    model = MultiTaskNLPModel.load(model_dir, encoder_name).float().to(device)
     model.eval()
     tokenizer = AutoTokenizer.from_pretrained(model_dir)
